Set AUTOBUILD_VCS_BRANCH: secondlife/viewer-build-util/which-branch
For a tag build used to release a package, github reports the tag name rather
than the branch to which it corresponds. Look up the corresponding branch and
pass it to autobuild as AUTOBUILD_VCS_BRANCH, which will set the vcs_branch
field in the package's autobuild-package.xml's package_description.

Otherwise, vcs_branch is reported for every build of every package as "HEAD".