Unconditionally use the GitHub workflow run ID as build ID.
AUTOBUILD_BUILD_ID is captured in the package tarball's autobuild-package.xml
as build_id. No other information in autobuild-package.xml permits uniquely
identifying the build that constructed the package in question. The changeset
hash could reference any of a number of builds of the same changeset.

Without knowing the specific build, we don't even know the date the package
was built.