Upgrade to actions/upload-artifact@v4
This changeset upgrades the composit action to use upload-artifact@v4,
which could be considered a breaking change:

https://github.com/actions/upload-artifact?tab=readme-ov-file#breaking-changes

Specifically, artifact upload is now immutable and will fail if an
artifact is uploading using the name of one that is already created.

Fortunately, this shouldn't be a problem for most consumers of
the action-autobuild package, as the platfom name is baked into
autobuild package and resulting artifact names.